fix: update CLAUDE.md pre-commit checks and fix clippy warnings
Align CLAUDE.md pre-commit checklist with CI commands. Fix clippy
warnings: unnecessary cast, collapsible if statements, redundant deref.

+++ crates/telegram/src/bot.rs
@@ -101,7 +101,7 @@ pub async fn start_polling(
 
             let result = bot
                 .get_updates()
-                .offset(offset as i32)
+                .offset(offset)
                 .timeout(30)
                 .allowed_updates(vec![AllowedUpdate::Message, AllowedUpdate::CallbackQuery])
                 .await;
